requirements.txt
tensorflow==2.12.0
numpy<2.0
opencv-python
pillow
matplotlib
scikit-learn
pandas
keras
